+++ sql/catalyst/src/main/scala/org/apache/spark/sql/catalyst/types/dataTypes.scala
@@ -572,7 +572,7 @@ case class MapType(
  * The data type for User Defined Types.
  */
 @DeveloperApi
-abstract class UserDefinedTypeType[UserType] extends DataType with Serializable {
+abstract class UserDefinedType[UserType] extends DataType with Serializable {
 
   /** Underlying storage type for this UDT used by SparkSQL */
   def sqlType: DataType
